Updated CharacterController2D
Added base assets for a teleporter and a bouncer.
Added the ability to jump multiple times up to a set amount before having to touch ground.
Kept the ability to hold down the jump key to chain jumps.
Added the ability to set the delay between consecutive jumps.
Renamed variables to make more sense on first read.
Added the ability to crouch. -> Still needs the crouch animation to be created.
Removed the jitter that happens when the player stands between two colliders at the same time.

Signed-off-by: h1ddengames <skarim3000@gmail.com>