Updated Unity Version, Added GameManager, Teleporter
Unity Version updated to Unity 2019.3.2f1

Character Controller now has an option to ignore gravity.
GameVersion keeps track of the version to display on the bottom left of the UI.
GameManager keeps track of various details about the game state such as the player and location of certain items/enemies/teleporters.
Teleporter moves the player towards another teleporter after turning off the sprite, and ignoring gravity. In order to make teleporters work, a rigidbody must be on either the player or the teleporter and be set to kinematic.

Signed-off-by: h1ddengames <skarim3000@gmail.com>