Added Fluent Automated Tasks
Tasks (Actions) can now be run with a delay before, an action before, the action itself, an action after, and finally a delay after.
These tasks can be saved as a ScriptableObject for easy reuse or they can be called through code during runtime.

The actual calling of this task chain has not been implemented yet.
The calling will utilize ScriptableObject cloning in order to implement the timer.
example:
	1. Game will start.
	2. ScriptableObject AutomatedTask will be cloned.
	3. The cloned ScriptableObject will have it's before timer decremented by Time.deltaTime.
	4. The before, during, after action will run.
	5. The after timer will be decremented.
	6. The cloned Scriptableobject will be disposed of.
	7. If the AutomatedTask list needs to be looped, restart the loop.

Signed-off-by: h1ddengames <skarim3000@gmail.com>