Added Unity's new Input System, Updated Render Pipeline
Currently not using Unity's new Input System due to the fact that documentation isn't complete and that the solution doesn't work for my purposes.

Created a PlayerInput script that manages keyboard inputs such as single press, double press, and long press (holding)
This PlayerInput class will be a Singleton that can be used by any other script in the scene.

Signed-off-by: h1ddengames <skarim3000@gmail.com>