Added a console system
Input is taken from the player through a TMP Input Field.
Commands have a command name and a UnityEvent.
When the player enters some text into the Input Field then presses enter, the console system checks all the available command names and if there's a match, the actions contained in the UnityEvent of that command will invoke.

Signed-off-by: h1ddengames <skarim3000@gmail.com>